#860: Merged count and scalarInt methods  into scalarLong.

+++ modules/core-domain/src/main/java/ccc/persistence/CommentRepository.java
@@ -102,5 +102,5 @@ public interface CommentRepository {
      *
      * @return The number of comments.
      */
-    int count(Resource resource, CommentStatus status);
+    long count(Resource resource, CommentStatus status);
 }
